Fix xp function unpickling
